Type Validator
Infrastructure:
"Type Validator" class validates the type of passing parameter to functions.
When the type is not as expected, it raises an exception.

Usage:
"Type Validator" class is invoked by functions to validate parameters, that are used inside the function.
Parameters that are not in use inside the current function, are ignored.